+++ minidump/test/minidump_rva_list_test_util.cc
@@ -26,7 +26,8 @@ namespace test {
 const MinidumpRVAList* MinidumpRVAListAtStart(const std::string& file_contents,
                                               size_t count) {
   MINIDUMP_LOCATION_DESCRIPTOR location_descriptor;
-  location_descriptor.DataSize = sizeof(MinidumpRVAList) + count * sizeof(RVA);
+  location_descriptor.DataSize =
+      static_cast<uint32_t>(sizeof(MinidumpRVAList) + count * sizeof(RVA));
   location_descriptor.Rva = 0;
 
   const MinidumpRVAList* list =
